+++ hrk_nasa_apis/lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/sbdb_cad_query_parameters.dart
@@ -62,6 +62,7 @@ class SbdbCadQueryParameters with _$SbdbCadQueryParameters {
     unit: distUnitDefault,
   );
   static const VelocityUnit velocityUnitDefault = VelocityUnit.kmps;
+  static const DistanceUnit diameterUnitDefault = DistanceUnit.km;
   static const SmallBodyFilter smallBodyFilterDefault = SmallBodyFilter.neo;
   static const CloseApproachBody closeApproachBodyDefault =
       CloseApproachBody.earth;