+++ hrk_nasa_apis/test/services/ssd_cneos/sbdb_cad_api/data/sbdb_cad_data_test.dart
@@ -20,6 +20,10 @@ void main() {
       test('Basic', () {
         testConverter(hk2023SbdbCadData, hk2023DataJson);
       });
+
+      test('All Fields', () {
+        testConverter(wnFive2001SbdbCadData, wnFive2001DataJson);
+      });
     });
   });
 }